Add support for extended patterns
This PR extend the grammar with the latest pattern developments in
upstream Nickel:

- constant patterns
- wildcard patterns
- array patterns
- or patterns
- pattern guards

Unfortunately, it's not trivial to properly handle identifier named `or`
inside patterns (which is indeed supported in the main parser of
Nickel). This is left for future work, because the problem needs a bit
more investigation to decide which is the best way to go in tree-sitter.