Simplify enum_variant_pattern
Instead of instantiating `patternF` manually, we can actually reuse the
rule `pattern_fun` within the `enum_variant_pattern`, which makes things
a bit more consistent and will help on the Topiary formatting side.

Passing by, we also removed the unused last parameter passed to the
various xxxF macro rules - this was copied from LALRPOP but doesn't make
sense currently in the tree-sitter grammar. Fortunately (or not?)
additional arguments are just ignored in JavaScript, so this shouldn't
change anything in practice, but is less confusing to future readers of
the grammar.